+++ src/Message/DataStreamInterface.php
@@ -1,0 +1,9 @@
+<?php
+namespace Pyncer\Http\Message;
+
+use Psr\Http\Message\StreamInterface as PsrStreamInterface;
+
+interface DataStreamInterface extends PsrStreamInterface
+{
+    public function getContentType(): string;
+}